fix: created role-request type to match backend for hooks
