+++ app/frontend/src/components/admin/approvalTable.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { RoleRequest, RoleStatus } from "../../types/admin";
+import { RoleRequest, RoleStatus } from "../../types/role-request";
 import { statusBadge, BadgeStyle } from "./statusBadge";
 
 interface RoleRequestTableProps {
